move DataInterface type and type guards to common
